+++ changelog.md
@@ -7,6 +7,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Add a `--config`/`-C` option to load CLI configuration from a TOML file.
+
 ## [1.0.1 (2021-10-27)](https://github.com/kdeldycke/click-extra/compare/v1.0.0...v1.0.1)
 
 - Re-release previous version with fixed dependency.